+++ src/Action/PortalExtension/PortalExtensionFind.php
@@ -14,7 +14,7 @@ use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Id;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryFactory;
 
-class PortalExtensionFind implements PortalExtensionFindActionInterface
+final class PortalExtensionFind implements PortalExtensionFindActionInterface
 {
     public const LOOKUP_QUERY = '82bb12c6-ed9c-4646-901a-4ff7e8e4e88c';