+++ src/Action/PortalNodeConfiguration/PortalNodeConfigurationSet.php
@@ -15,7 +15,7 @@ use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Id;
 
-class PortalNodeConfigurationSet implements PortalNodeConfigurationSetActionInterface
+final class PortalNodeConfigurationSet implements PortalNodeConfigurationSetActionInterface
 {
     private Connection $connection;