+++ src/StorageKey/PortalNodeStorageKey.php
@@ -7,7 +7,7 @@ namespace Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\AliasAwarePortalNodeStorageKey;
 
-class PortalNodeStorageKey extends AbstractStorageKey implements PortalNodeKeyInterface
+final class PortalNodeStorageKey extends AbstractStorageKey implements PortalNodeKeyInterface
 {
     public function withAlias(): PortalNodeKeyInterface
     {